client state: Add functions to add/remove reducers
This commit adds functionality to add and remove reducers dynamically.
This code isn't used yet in the combined reducer as that is the next
step.

The purpose for this code is to be able to dynamically add and
eventually remove Calypso extensions from the running app. This is
necessary because the extentions are in their own JS bundles and load as
needed, so they need to be able to be included in the reducer chain
only when they're loaded.