PostSchedule: Add AM/PM selector
This PR adds an AM/PM selector to the scheduling component of the
editor.

Fixes #10762.

To test: Make sure your site’s time format is set to include AM/PM
(setting is still in wp-admin). Navigate to the editor and open the
calendar next to the Publish button. Time should adjust in the
background.