Extensions: Move Hello World markup to own file.
The index.js of the example plugin was already a mix of application
logic and markup. This is going to become moreso as we add state and
other capabilities to the extensions. So this is an effort to simplify
the code and split out the display from the main application logic of
the extension.